+++ messages/en.json
@@ -790,17 +790,18 @@
     "warn_view_syntax": "Usage: warn view <id>",
     "warn_ack_syntax": "Usage: warn ack <id>",
     "warn_help_syntax": "Usage: warn help <subcommand>",
-    "fwarn_add_syntax": "Usage: fwarn add <nick[!user@host]|=account> [@]<points> [~expiry] [sanctions] <:reason> [| notes]",
+    "fwarn_add_syntax": "Usage: fwarn add <nick[!user@host]|=account> <points> [~expiry] [sanctions] [:]<reason> [| notes]",
     "fwarn_page_invalid": "Invalid page, must be a number 1 or greater.",
     "fwarn_points_invalid": "Invalid points, must be a number above 0.",
     "fwarn_expiry_invalid": "Invalid expiration, must be a number above 0 followed by either d, h, or m, or 'never' for a warning that never expires.",
     "fwarn_cannot_add": "Cannot add warning, double-check your parameters (the nick might be wrong or you are not joined to the channel).",
     "fwarn_added": "Added warning #{0}.",
     "fwarn_done": "Done.",
-    "fwarn_sanction_invalid": "Invalid sanction, can be either deny or stasis.",
+    "fwarn_sanction_invalid": "Invalid sanction, can be either deny, stasis, or tempban.",
     "fwarn_stasis_invalid": "Invalid stasis amount, specify sanction as \"stasis=number\".",
     "fwarn_deny_invalid": "Invalid denied commands, specify sanction as \"deny=command,command,command\" (without spaces).",
     "fwarn_deny_invalid_command": "Invalid command \"{0}\", specify sanction as \"deny=command,command,command\" (without spaces).",
+    "fwarn_tempban_invalid": "Invalid tempban amount, specify sanction as \"tempban=number\" or \"tempban=expiration\" (followed by d, h, or m).",
     "fwarn_list_header": "{0} has {1} active warning point{2}. Warnings prefixed with \u0002!\u0002 are unacknowledged.",
     "warn_list_header": "You have {0} active warning point{1}. You must acknowledge all warnings prefixed with \u0002!\u0002 by using \"warn ack <id>\" before you can join games.",
     "fwarn_list": "{0}{1}[#{2} {3}] to {4} by {5} - {6} ({7} point{8}, {9}){10}",
@@ -825,6 +826,7 @@
     "fwarn_view_stasis_sing": "1 game of stasis.",
     "fwarn_view_stasis_plural": "{0} games of stasis.",
     "fwarn_view_deny": "denied {0}.",
+    "fwarn_view_tempban": "banned until {0} (if a number, indicates warning point threshold).",
     "fwarn_reason_required": "A public warning reason is required.",
     "warn_unacked": "You have unacknowledged warnings and cannot join at this time. Use \"warn list\" to view them.",
     "no_templates": "There are no access templates defined.",
@@ -846,6 +848,7 @@
     "part_warning": "Parting during game. You only have a short time to re-join the channel after parting to stay alive.",
     "quit_warning": "Quitting IRC during game. You only have a short time to re-join the channel after quitting to stay alive.",
     "acc_warning": "Changing accounts during game. Please do not change accounts while playing.",
+    "tempban_kick": "{nick}",
 
     "_": " vim: set sw=4 expandtab:"
 }
