+++ src/db/db.sql
@@ -4,7 +4,7 @@
 
 -- Player tracking. This is just what the bot decides is a unique player, two entries
 -- here may end up corresponding to the same actual person (see below).
-CREATE TABLE IF NOT EXISTS player (
+CREATE TABLE player (
     id INTEGER PRIMARY KEY,
     -- What person this player record belongs to
     person INTEGER REFERENCES person(id) DEFERRABLE INITIALLY DEFERRED,
@@ -17,13 +17,13 @@ CREATE TABLE IF NOT EXISTS player (
     active BOOLEAN NOT NULL DEFAULT 1
 );
 
-CREATE INDEX IF NOT EXISTS player_idx ON player (account, hostmask, active);
-CREATE INDEX IF NOT EXISTS person_idx ON player (person);
+CREATE INDEX player_idx ON player (account, hostmask, active);
+CREATE INDEX person_idx ON player (person);
 
 -- Person tracking; a person can consist of multiple players (for example, someone may have
 -- an account player for when they are logged in and 3 hostmask players for when they are
 -- logged out depending on what connection they are using).
-CREATE TABLE IF NOT EXISTS person (
+CREATE TABLE person (
     id INTEGER PRIMARY KEY,
     -- Primary player for this person
     primary_player INTEGER NOT NULL UNIQUE REFERENCES player(id) DEFERRABLE INITIALLY DEFERRED,
@@ -45,7 +45,7 @@ CREATE TABLE IF NOT EXISTS person (
 -- Sometimes people are bad, this keeps track of that for the purpose of automatically applying
 -- various sanctions and viewing the past history of someone. Outside of specifically-marked
 -- fields, records are never modified or deleted from this table once inserted.
-CREATE TABLE IF NOT EXISTS warning (
+CREATE TABLE warning (
     id INTEGER PRIMARY KEY,
     -- The target (recipient) of the warning
     target INTEGER NOT NULL REFERENCES person(id) DEFERRABLE INITIALLY DEFERRED,
@@ -73,13 +73,13 @@ CREATE TABLE IF NOT EXISTS warning (
     deleted_on DATETIME
 );
 
-CREATE INDEX IF NOT EXISTS warning_idx ON warning (target, deleted, issued);
-CREATE INDEX IF NOT EXISTS warning_sender_idx ON warning (target, sender, deleted, issued);
+CREATE INDEX warning_idx ON warning (target, deleted, issued);
+CREATE INDEX warning_sender_idx ON warning (target, sender, deleted, issued);
 
 -- In addition to giving warning points, a warning may have specific sanctions attached
 -- that apply until the warning expires; for example preventing a user from joining deadchat
 -- or denying them access to a particular command (such as !goat).
-CREATE TABLE IF NOT EXISTS warning_sanction (
+CREATE TABLE warning_sanction (
     -- The warning this sanction is attached to
     warning INTEGER NOT NULL REFERENCES warning(id) DEFERRABLE INITIALLY DEFERRED,
     -- The type of sanction this is
@@ -92,7 +92,7 @@ CREATE TABLE IF NOT EXISTS warning_sanction (
 -- This shouldn't be too horribly slow, but if it is some strategies can be employed to speed it up:
 -- On startup, aggregate everything from this table and store in-memory, then increment those in-memory
 -- counts as games are played.
-CREATE TABLE IF NOT EXISTS game (
+CREATE TABLE game (
     id INTEGER PRIMARY KEY,
     -- The gamemode played
     gamemode TEXT NOT NULL COLLATE NOCASE,
@@ -110,10 +110,10 @@ CREATE TABLE IF NOT EXISTS game (
     winner TEXT COLLATE NOCASE
 );
 
-CREATE INDEX IF NOT EXISTS game_idx ON game (gamemode, gamesize);
+CREATE INDEX game_idx ON game (gamemode, gamesize);
 
 -- List of people who played in each game
-CREATE TABLE IF NOT EXISTS game_player (
+CREATE TABLE game_player (
     id INTEGER PRIMARY KEY,
     game INTEGER NOT NULL REFERENCES game(id) DEFERRABLE INITIALLY DEFERRED,
     player INTEGER NOT NULL REFERENCES player(id) DEFERRABLE INITIALLY DEFERRED,
@@ -125,11 +125,11 @@ CREATE TABLE IF NOT EXISTS game_player (
     dced BOOLEAN NOT NULL
 );
 
-CREATE INDEX IF NOT EXISTS game_player_game_idx ON game_player (game);
-CREATE INDEX IF NOT EXISTS game_player_player_idx ON game_player (player);
+CREATE INDEX game_player_game_idx ON game_player (game);
+CREATE INDEX game_player_player_idx ON game_player (player);
 
 -- List of all roles and other special qualities (e.g. lover, entranced, etc.) the player had in game
-CREATE TABLE IF NOT EXISTS game_player_role (
+CREATE TABLE game_player_role (
     game_player INTEGER NOT NULL REFERENCES game_player(id) DEFERRABLE INITIALLY DEFERRED,
     -- Name of the role or other quality recorded
     role TEXT NOT NULL COLLATE NOCASE,
@@ -137,11 +137,11 @@ CREATE TABLE IF NOT EXISTS game_player_role (
     special BOOLEAN NOT NULL
 );
 
-CREATE INDEX IF NOT EXISTS game_player_role_idx ON game_player_role (game_player);
+CREATE INDEX game_player_role_idx ON game_player_role (game_player);
 
 -- Access templates; instead of manually specifying flags, a template can be used to add a group of
 -- flags simultaneously.
-CREATE TABLE IF NOT EXISTS access_template (
+CREATE TABLE access_template (
 	id INTEGER PRIMARY KEY,
 	-- Template name, for display purposes
 	name TEXT NOT NULL,
@@ -150,7 +150,7 @@ CREATE TABLE IF NOT EXISTS access_template (
 );
 
 -- Access control, owners still need to be specified in botconfig, but everyone else goes here
-CREATE TABLE IF NOT EXISTS access (
+CREATE TABLE access (
 	person INTEGER NOT NULL PRIMARY KEY REFERENCES person(id) DEFERRABLE INITIALLY DEFERRED,
 	-- Template to base this person's access on, or NULL if it is not based on a template
 	template INTEGER REFERENCES access_template(id) DEFERRABLE INITIALLY DEFERRED,
@@ -159,8 +159,15 @@ CREATE TABLE IF NOT EXISTS access (
 	flags TEXT
 );
 
+-- Holds bans that the bot is tracking (due to sanctions)
+CREATE TABLE bantrack (
+	player INTEGER NOT NULL PRIMARY KEY REFERENCES player(id) DEFERRABLE INITIALLY DEFERRED,
+	expires DATETIME,
+	warning_amount INTEGER
+);
+
 -- Used to hold state between restarts
-CREATE TABLE IF NOT EXISTS pre_restart_state (
+CREATE TABLE pre_restart_state (
 	-- List of players to ping after the bot comes back online
 	players TEXT
 );
